backend/payments/services.py
"""
Services de paiement pour Bankily, Moov Money, BAMIS, BMCI et cartes bancaires
"""
import requests
from django.conf import settings
from decouple import config


class PaymentService:
    """Service de base pour les paiements"""
    
    def __init__(self):
        self.api_key = None
        self.api_url = None
    
    def process_payment(self, amount, reference, user_info):
        """Traite un paiement"""
        raise NotImplementedError("Cette méthode doit être implémentée")
    
    def verify_payment(self, transaction_id):
        """Vérifie le statut d'un paiement"""
        raise NotImplementedError("Cette méthode doit être implémentée")


class BankilyService(PaymentService):
    """
    Service de paiement Bankily
    Note: Implémentation basique - nécessite les vraies clés API
    """
    
    def __init__(self):
        super().__init__()
        self.api_key = config('BANKILY_API_KEY', default='')
        self.api_url = config('BANKILY_API_URL', default='https://api.bankily.mr')
    
    def process_payment(self, amount, reference, user_info):
        """
        Traite un paiement via Bankily
        """
        # TODO: Implémenter l'intégration réelle avec l'API Bankily
        # Exemple de structure:
        payload = {
            'amount': str(amount),
            'reference': reference,
            'phone': user_info.get('telephone'),
            'callback_url': f"{settings.ALLOWED_HOSTS[0]}/api/payments/bankily/callback"
        }
        
        # headers = {
        #     'Authorization': f'Bearer {self.api_key}',
        #     'Content-Type': 'application/json'
        # }
        # response = requests.post(f"{self.api_url}/payment", json=payload, headers=headers)
        # return response.json()
        
        # Pour le développement, retourner une réponse simulée
        return {
            'success': True,
            'transaction_id': f'BANKILY_{reference}',
            'status': 'pending',
            'message': 'Paiement initié avec succès'
        }
    
    def verify_payment(self, transaction_id):
        """Vérifie le statut d'un paiement Bankily"""
        # TODO: Implémenter la vérification réelle
        return {
            'success': True,
            'status': 'completed',
            'transaction_id': transaction_id
        }


class MoovMoneyService(PaymentService):
    """
    Service de paiement Moov Money
    """
    
    def __init__(self):
        super().__init__()
        self.api_key = config('MOOV_MONEY_API_KEY', default='')
        self.api_url = config('MOOV_MONEY_API_URL', default='https://api.moov.mr')
    
    def process_payment(self, amount, reference, user_info):
        """
        Traite un paiement via Moov Money
        """
        # TODO: Implémenter l'intégration réelle avec l'API Moov Money
        return {
            'success': True,
            'transaction_id': f'MOOV_{reference}',
            'status': 'pending',
            'message': 'Paiement initié avec succès'
        }
    
    def verify_payment(self, transaction_id):
        """Vérifie le statut d'un paiement Moov Money"""
        return {
            'success': True,
            'status': 'completed',
            'transaction_id': transaction_id
        }


class BAMISService(PaymentService):
    """
    Service de paiement BAMIS
    """
    
    def __init__(self):
        super().__init__()
        self.api_key = config('BAMIS_API_KEY', default='')
        self.api_url = config('BAMIS_API_URL', default='https://api.bamis.mr')
    
    def process_payment(self, amount, reference, user_info):
        """
        Traite un paiement via BAMIS
        """
        # TODO: Implémenter l'intégration réelle avec l'API BAMIS
        return {
            'success': True,
            'transaction_id': f'BAMIS_{reference}',
            'status': 'pending',
            'message': 'Paiement initié avec succès'
        }
    
    def verify_payment(self, transaction_id):
        """Vérifie le statut d'un paiement BAMIS"""
        return {
            'success': True,
            'status': 'completed',
            'transaction_id': transaction_id
        }


class BMCIService(PaymentService):
    """
    Service de paiement BMCI
    """
    
    def __init__(self):
        super().__init__()
        self.api_key = config('BMCI_API_KEY', default='')
        self.api_url = config('BMCI_API_URL', default='https://api.bmci.mr')
    
    def process_payment(self, amount, reference, user_info):
        """
        Traite un paiement via BMCI
        """
        # TODO: Implémenter l'intégration réelle avec l'API BMCI
        return {
            'success': True,
            'transaction_id': f'BMCI_{reference}',
            'status': 'pending',
            'message': 'Paiement initié avec succès'
        }
    
    def verify_payment(self, transaction_id):
        """Vérifie le statut d'un paiement BMCI"""
        return {
            'success': True,
            'status': 'completed',
            'transaction_id': transaction_id
        }


class CarteBancaireService(PaymentService):
    """
    Service de paiement par carte bancaire (Visa/Mastercard)
    Utilise Stripe ou un PSP local (GIMTEL)
    """
    
    def __init__(self):
        super().__init__()
        import stripe
        self.stripe_key = config('STRIPE_SECRET_KEY', default='')
        if self.stripe_key:
            stripe.api_key = self.stripe_key
        self.stripe = stripe
    
    def process_payment(self, amount, reference, user_info):
        """
        Traite un paiement par carte bancaire
        """
        # TODO: Implémenter l'intégration réelle avec Stripe ou GIMTEL
        # Exemple avec Stripe:
        # try:
        #     payment_intent = self.stripe.PaymentIntent.create(
        #         amount=int(amount * 100),  # Convertir en centimes
        #         currency='mru',
        #         metadata={'reference': reference}
        #     )
        #     return {
        #         'success': True,
        #         'transaction_id': payment_intent.id,
        #         'client_secret': payment_intent.client_secret,
        #         'status': 'pending'
        #     }
        # except Exception as e:
        #     return {
        #         'success': False,
        #         'error': str(e)
        #     }
        
        return {
            'success': True,
            'transaction_id': f'CARTE_{reference}',
            'status': 'pending',
            'message': 'Paiement initié avec succès'
        }
    
    def verify_payment(self, transaction_id):
        """Vérifie le statut d'un paiement par carte"""
        # TODO: Implémenter la vérification réelle
        return {
            'success': True,
            'status': 'completed',
            'transaction_id': transaction_id
        }


def get_payment_service(mode_paiement):
    """
    Retourne le service de paiement approprié selon le mode
    """
    services = {
        'bankily': BankilyService,
        'moov_money': MoovMoneyService,
        'bamis': BAMISService,
        'bmci': BMCIService,
        'carte': CarteBancaireService,
    }
    
    service_class = services.get(mode_paiement)
    if not service_class:
        raise ValueError(f"Mode de paiement non supporté: {mode_paiement}")
    
    return service_class()

