backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.11
#Pillow==10.1.0
python-decouple==3.8
django-filter==23.5
celery==5.3.4
redis==5.0.1
django-allauth==0.54.0
stripe==7.8.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0

